% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareDataforGlm.R
\name{analyzeSeparation}
\alias{analyzeSeparation}
\title{Find a hyperplane separating the sample points}
\usage{
analyzeSeparation(data, tol1 = 0.001, tol2 = 1e-09)
}
\arguments{
\item{data}{A data.frame containing a column called Response, holding a
binary vector.}

\item{tol1}{The tolerance for considering the points separated: the sum of
all betas has to be larger than tol1.}

\item{tol2}{The tolerance for considering an individual feature to be
            involved.}
}
\value{
A list with two slots: a logical value saying if the data points are
        separated and a vector of distances to the separating hyperplane.
}
\description{
Given a data frame holding feature data and a binary response column, find a
hyperplane separating the two groups of data points. The core code solving
the linear program is taken from the package safeBinaryRegression.
}
\examples{
# get gene locations
mtor.loc <- findWells(experiments="brucella-du-k1", contents="MTOR")
scra.loc <- findWells(plates=sapply(mtor.loc, getBarcode),
                      contents="SCRAMBLED", well.names="G23")
# combine for faster fetching
data     <- getSingleCellData(list(mtor.loc[[1]], scra.loc[[1]]))
mtor.dat <- meltData(cleanData(data[[1]]$H6))
scra.dat <- meltData(cleanData(data[[1]]$G23))
# prepare data for glm
data <- prepareDataforGlm(mtor.dat$mat$Cells, scra.dat$mat$Cells)
data <- makeRankFull(data)
# run glm
model <- glm("Response ~ .", binomial, data$train)
# compare to testing data
predi <- as.factor(round(predict(model, newdata=data$test,
                                 type="response")))
levels(predi) <- c("active", "control")
comparison <- compareModeltoTruth(predi, data$test$Response)
}

