% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augmentAggregate.R
\name{augmentAggregate}
\alias{augmentAggregate}
\title{Augment PlateData with Aggregates}
\usage{
augmentAggregate(x, ...)
}
\arguments{
\item{x}{The PlateData object of interest.}

\item{level}{A string, either "well" or "plate", indicaing the level
at which to aggregate.}

\item{neighbors}{Logical, whether to include direct neighbors for well level
aggregation.}

\item{features}{A (list) of regular expressions used for selecting the
features to include.}

\item{drop}{A (list) of regular expressions used for removing features
from the previously compiled list.}

\item{aggregate}{The function used to aggregate data within wells.}
}
\value{
An augmented version of the input object.
}
\description{
Given a PlateData, aggregate the selected features within wells or the
whole plate with the supplied function (examples include mean, median, sum,
var, mad, etc.)
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
aug  <- augmentAggregate(data)
}

