% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augmentBscore.R
\name{augmentBscore}
\alias{augmentBscore}
\title{Augment PlateData with B-scores}
\usage{
augmentBscore(x, ...)
}
\arguments{
\item{x}{The PlateData object of interest.}

\item{features}{A regular expression used for selecting the features to
include.}

\item{aggregate}{The function used to aggregate data within wells.}
}
\value{
An augmented version of the input object.
}
\description{
Given a PlateData, aggregate the selected features within wells and
calculate their plate B-scores.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
aug  <- augmentBscore(data)
}

