% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augmentImageLocation.R
\name{augmentImageLocation}
\alias{augmentImageLocation}
\title{Augment the image location within a well}
\usage{
augmentImageLocation(x)
}
\arguments{
\item{x}{The PlateData/WellData/ImageData object of interest.}
}
\value{
An augmented version of the input object.
}
\description{
Given a PlateData/WellData/ImageData object, add the "Image.Group" feature
which encodes how many well borders are image borders and for every
"Location_Center_X/Y" feature, add "Location_Shifted_X/Y" feature for where
the object is located within the well.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
aug <- augmentImageLocation(data)
}

