% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augmentMars.R
\name{augmentMars}
\alias{augmentMars}
\title{Augment PlateData with MARS residuals}
\usage{
augmentMars(x, ...)
}
\arguments{
\item{x}{The PlateData object of interest.}

\item{bscore}{Whether to include B-scores (logical).}

\item{model}{A (list) of regular expressions, specifying the features
used for MARS modeling.}

\item{features}{A (list) of regular expressions used for selecting the
features to include.}

\item{drop}{A (list) of regular expressions used for removing features
                 from the previously compiled list.}
}
\value{
An augmented version of the input object.
}
\description{
Given a PlateData, normalize the selected features with a MARS model
consisting of terms for membership to image/well centered ellipses and
medpolish estimated row/column effects.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
data <- augmentImageLocation(data)
data <- augmentCordinateFeatures(data, 1, NULL, FALSE, FALSE)
data <- augmentBscore(data)
data <- augmentMars(data)
}

