% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkCompletenessImage.R
\name{checkCompletenessImage}
\alias{checkCompletenessImage}
\title{Check if the set of images is complete}
\usage{
checkCompletenessImage(...)
}
\arguments{
\item{x}{The object of interest.}
}
\value{
Logical: TRUE if complete, FALSE if incomplete.
}
\description{
Given a PlateData or WellData object, check if the set of images is complete
over the entire object.
}
\examples{
plate <- PlateLocation("J101-2C")
checkCompletenessImage(plate)
}

