% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkConsistency.R
\name{checkConsistency}
\alias{checkConsistency}
\title{Check consistency of Data object}
\usage{
checkConsistency(...)
}
\arguments{
\item{x}{The object of interest.}
}
\value{
Logical: TRUE if consistent, FALSE if inconsistent (the
inconsistencies are printed).
}
\description{
Given a PlateData or WellData object, check if all sub-objects are
consistent, i.e. if they contain the same features, barcodes, well names.
}
\examples{
plate <- PlateLocation("J101-2C")
checkConsistency(plate)
}

