% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkDataCompleteness.R
\name{checkDataCompleteness}
\alias{checkDataCompleteness}
\title{Check completeness of data}
\usage{
checkDataCompleteness(x)
}
\arguments{
\item{x}{Data object}
}
\value{
A list holding two vectors of strings, one for features that are
        expected but missing and one for features that are present but not
        expected.
}
\description{
Given a Data object (currently only MatData objects supported), determine if
the set of available features is complete.
}
\examples{
data  <- MatData(PlateLocation("J107-2C"))
checkDataCompleteness(data)
}

