% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkFeatureCompleteness.R
\name{checkFeatureCompleteness}
\alias{checkFeatureCompleteness}
\title{Check completeness of data}
\usage{
checkFeatureCompleteness(x)
}
\arguments{
\item{x}{Data object}
}
\value{
A list holding two vectors of strings, one for features that are
        expected but missing and one for features that are present but not
        expected.
}
\description{
Given a Data object, determine if the set of available features is complete.
}
\examples{
data  <- MatData(PlateLocation("J107-2C"))
checkFeatureCompleteness(data)
}

