% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Remove some bad image data}
\usage{
cleanData(x, ...)
}
\arguments{
\item{x}{The PlateData/WellData object of interest.}

\item{quants}{A string for choosing between only applying lower, upper,
both one none of the quantiles for cleaning up the data.}

\item{infected}{A string for choosing whether to discard infected or
uninfected cells.}
}
\value{
A cleaned up version of the input object.
}
\description{
Given a PlateData/WellData object, remove all images where the cell count
does not lie within the 0.05 and 0.95 quantiles.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
cleaned <- cleanData(data)

}

