% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{configGet}
\alias{configGet}
\title{Loads the config object}
\usage{
configGet()
}
\value{
The config object.
}
\description{
Loads the yaml file saved at the current config file location.
}
\examples{
config <- configGet()
}

