% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{configInit}
\alias{configInit}
\title{Creates a config file template}
\usage{
configInit()
}
\value{
NULL (invisibly)
}
\description{
Saves a config file template to the current config file path.
}
\examples{
configInit()

}

