% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{configPathGet}
\alias{configPathGet}
\title{Get the config file path}
\usage{
configPathGet()
}
\value{
The config file path
}
\description{
Gets the path to the yaml configuration file for the singleCellFeatures
package.
}
\examples{
path <- configPathGet()
}

