% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{configPathSet}
\alias{configPathSet}
\title{Set the config file path}
\usage{
configPathSet(path)
}
\arguments{
\item{path}{The new path of the config file.}
}
\value{
NULL (invisibly)
}
\description{
Sets the path to the yaml configuration file for the singleCellFeatures
package.
}
\examples{
configPathSet("path/to/config.yaml")
}

