% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{configSet}
\alias{configSet}
\title{Overwrites the config file}
\usage{
configSet(config)
}
\arguments{
\item{x}{The new config object.}
}
\value{
NULL (invisibly)
}
\description{
Saves the specified config object to the current config file location. Yaml
was chosen as config file format due to its human read/writability, so direct
editing of the file is possible. This function is available for convenience
nevertheless.
}
\examples{
conf <- configGet()
# edit the list in R
configSet(conf)

}

