% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToPlateLocation.R
\name{convertToPlateLocation}
\alias{convertToPlateLocation}
\title{Covert compatible objects to PlateLocation}
\usage{
convertToPlateLocation(x)
}
\arguments{
\item{The}{object to be used as basis for new PlateLocation object}
}
\value{
The new PlateLocation object
}
\description{
Build PlateLocation object from other objects, such as PlateData,
MatData or WellLocation
}
\examples{
plate <- PlateLocation("J101-2C")
path <- convertToPlateLocation(plate)

}

