% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convertToWellLocation.R
\name{convertToWellLocation}
\alias{convertToWellLocation}
\title{Covert compatible objects to WellLocation}
\usage{
convertToWellLocation(x)
}
\arguments{
\item{The}{object to be used as basis for new WellLocation object}
}
\value{
The new WellLocation object
}
\description{
Build WellLocation object from other objects, such as WellData or
WellMetadata
}
\examples{
plate <- PlateLocation("J101-2C")
path <- convertToWellLocation(plate)
}

