% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilLocalData.R
\name{deleteAllMetadataCaches}
\alias{deleteAllMetadataCaches}
\title{Remove all metadata caches}
\usage{
deleteAllMetadataCaches()
}
\value{
NULL (invisibly).
}
\description{
Due to generating metadata caches being fairly inexpensive and the
circumstance that they are downstream to Data object creation (changes in
Data/Metadata object specifications often require all metadata caches to be
rebuilt), this function purges all metadata caches.
}
\examples{
deleteAllWellCaches()

}

