% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilLocalData.R
\name{deleteAllWellCaches}
\alias{deleteAllWellCaches}
\title{Remove all local well caches}
\usage{
deleteAllWellCaches()
}
\value{
NULL (invisibly).
}
\description{
Due to generating well caches being fairly inexpensive and the circumstance
that they are downstream to Data object creation (changes in Data object
specifications often require all well caches to be rebuilt), this function
purges all well caches. All .rds files in well cache directories are deleted
and if the direcory is empty afterwards, it is also removed.
}
\examples{
deleteAllWellCaches()

}

