% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractCells.R
\name{extractCells}
\alias{extractCells}
\title{Extract cells from Data objects}
\usage{
extractCells(x, ...)
}
\arguments{
\item{x}{The original Data object.}

\item{infected}{Either NULL or a boolgen whether to return all infected (
                TRUE) or all uninfected cells (FALSE).}
}
\value{
A Data object of the same type as x, but only containing a subset of
        all cells.
}
\description{
Given a Data object, extract all infected or all uninfected cells.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
data <- extractCells(data, infected=TRUE)
}

