% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Extract features from Data objects}
\usage{
extractFeatures(x, select = NULL, drop = NULL, features = NULL)
}
\arguments{
\item{x}{The original Data object, containing the superset of
features}

\item{select}{A vector of strings (regular expressions), specifying
features to keep}

\item{drop}{A vector of strings (regular expressions), specifying
features to drop}

\item{features}{For performance reasons, the complete set of features to
keep can be determined in advance and specified here (in
this case, arguments select/drop will be ignored if not
NULL)}
}
\value{
A Data object of the same type as x, but only storing a subset of
        all features.
}
\description{
Given a Data object and either a set of regular expressions to select or
discard features or a set of features, return the same object type as the
original, but only with a subset of features.
}
\examples{
features <- c("^Cells.Location_Center_X$",
              "^Cells.Location_Center_Y$")
data <- PlateData(PlateLocation("J101-2C"))
data <- extractFeatures(data, features)

}

