% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractImages.R
\name{extractImages}
\alias{extractImages}
\title{Extract data for images from Data objects}
\usage{
extractImages(...)
}
\arguments{
\item{x}{The original Data object, holding the superset of images.}

\item{images}{A vector of integers, specifying image indices of the
images to keep.}

\item{keep.well}{(Only at WellData level) A boolean indicating whether to
keep the WellData object intact or return ImageData
object(s).}

\item{wells}{(Only at PlateData level) A list of wells (see 
extractWells), to further limit the returned data set.}

\item{keep.plate}{(Only at PlateData level) A boolean indicating whether to
keep the PlateData object intact or return a (nested) list.}
}
\value{
Either a PlateData, WellData or ImageData object, or a possibly
        nested list, holding only the data corresponding to the specified
        subset of images.
}
\description{
Given a Data object and a vector of image indices (either in 1:6 or in 1:9)
return only the data belonging to the selected images. If this extraction is
performed on well level, either a well with only a subset of images is
returned or a single ImageData object (or a list of ImageData) objects is/are
returned.
}
\examples{
plate  <- PlateData(PlateLocation("J101-2C"))
img_11 <- extractImages(plate, 1)
imgs   <- extractImages(plate, c(5, 6), c("A15", "P2"))

}

