% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extractPartialPlate.R
\name{extractPartialPlate}
\alias{extractPartialPlate}
\title{Given feature data from a whole plate, extract the data for specified rows/
cols/wells/images}
\usage{
extractPartialPlate(data, rows = NULL, cols = NULL, wels = NULL,
  imgs = 5)
}
\arguments{
\item{data}{Data for a whole plate, multiply nested list of features}

\item{rows}{Row(s) of the well(s), specified either by a vector consisting
of the characters A:P, a:P, or 1:16}

\item{cols}{Column(s) of the well(s), specified by a vector of integers in
the range of 1:24}

\item{wels}{Well index(es), specified by a vector of integers in the range
of 1:384}

\item{imgs}{Image number(s) within the well, specified by a vector of
            integers in the range of 1:9}
}
\value{
A multiply nested list of features with all values for the given
        plate coordinates (for structure, refer to importCompletePlate)
}
\description{
Given feature data from a whole plate, extract the data for specified rows/
cols/wells/images
}
\examples{
subset.img <- extractPartialPlate(dat.plat)
subset.wel <- extractPartialPlate(dat.plat, cols=c(3,4), rows=c("A","B"),
                                  wels=c(333, 21), imgs=NULL)
}

