% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWells.R
\name{extractWells}
\alias{extractWells}
\title{Extract WellData objects form a PlateData object}
\usage{
extractWells(x, ...)
}
\arguments{
\item{x}{The original PlateData object, holding the superset of
wells}

\item{wells}{A single WellLocation object/a list of WellLocation
objects, a single integer/vector of integers (in 1:384),
or a single string/vector of strings (e.g. A5) specifying
the targets.}

\item{keep.plate}{A boolean indicating whether to keep the PlateData object
intact or return WellData object(s)}
}
\value{
Either a PlateData or a WellData object holding only the data
        corresponding to the specified subset of wells
}
\description{
Given a PlateData object and a single WellLocation object of a list of
WellLocation objects, return either a PlateData object containing only the 
specified subset of wells or a single WellData/a list of WellData objects.
}
\examples{
plate  <- PlateData(PlateLocation("J101-2C"))
a23 <- extractWells(plate, "A23", FALSE)
wells <- extractWells(plate, c(135, 155))

}

