% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findDataLocations.R
\name{findPlates}
\alias{findPlates}
\title{Find PlateLocations}
\usage{
findPlates(pathogens = NULL, experiments = NULL, plates = NULL,
  well.types = NULL, contents = NULL, id.openBIS = NULL, id.infx = NULL,
  name = NULL, verbose = FALSE)
}
\arguments{
\item{pathogens}{Narrow down the results by selecting pathogens
(exact, matches, ignoring case).}

\item{experiments}{Name of the experiment to be considered (regular
expression, case insensitive).}

\item{plates}{A vector of regular expressions for plates.}

\item{well.types}{A vector of regular expressions that is matched with
the well.types column in well databases.}

\item{contents}{This vector of strings/integers is matched against
the three well database columns id.openBIS,
id.infx and name via case insensitive regular
expressions. In case it is specified, the next three
parameters are ignored.}

\item{id.openBIS}{A vector of regular expressions that is matched
against the id.openBIS column.}

\item{id.infx}{A vector of regular expressions that is matched
against the id.infx column.}

\item{name}{A vector of regular expressions that is matched
                       against the name column.}
}
\value{
A list of PlateLocation objects.
}
\description{
Find all plates using the following set of parameters:
}
\examples{
set1 <- findPlates(contents="MTOR", experiments="brucella-du-k")
set2 <- findPlates(contents=2475, experiments="brucella-au-k[1-3]")
set3 <- findPlates(contents="SCRAMBLED", experiments="brucella-du-k")
}

