% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findPlateWellsFromGene.R
\name{findPlateWellsFromGene}
\alias{findPlateWellsFromGene}
\title{Find wells corresponding to a gene}
\usage{
findPlateWellsFromGene(experiment, gene)
}
\arguments{
\item{experiment}{Name of the experiment to be considered (should start with
pathogen name, all converted to uppercase)
to be imported}

\item{gene}{String specifying the gene (case sensitive)}
}
\value{
A table containing columns for experiment, barcode, well row and well
        column
}
\description{
Find all wells on all plates of a given pathogen/experiment that contain
the specified gene
}
\examples{
mtor.bruc.du <- findPlateWellsFromGene("brucella-du-k", "MTOR")
scra.bruc.du <- findPlateWellsFromGene("brucella-du-k", "SCRAMBLED")
}

