% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findWells.R
\name{findWells}
\alias{findWells}
\title{Find wells}
\usage{
findWells(pathogens = NULL, experiments = NULL, plates = NULL,
  well.rows = NULL, well.cols = NULL, well.names = NULL,
  well.types = NULL, contents = NULL, id.openBIS = NULL, id.infx = NULL,
  name = NULL, verbose = FALSE)
}
\arguments{
\item{pathogens}{Narrow down the results by selecting pathogens
(exact, matches, ignoring case).}

\item{experiments}{Name of the experiment to be considered (regular
expression, case insensitive).}

\item{plates}{A vector of regular expressions for plates.}

\item{well.rows}{All rows to be included (a vector of characters,
exact match, case insensitive)}

\item{well.cols}{The columns to be included (a vector of integers)}

\item{well.names}{A vector of well names (e.g. B14), expecting strings,
case insensitive matching.}

\item{well.types}{A vector of regular expressions that is matched with
the well.types column in well databases.}

\item{contents}{This vector of strings/integers is matched against
the three well database columns id.openBIS,
id.infx and name via case insensitive regular
expressions. In case it is specified, the next three
parameters are ignored.}

\item{id.openBIS}{A vector of regular expressions that is matched
against the id.openBIS column.}

\item{id.infx}{A vector of regular expressions that is matched
against the id.infx column.}

\item{name}{A vector of regular expressions that is matched
                       against the name column.}
}
\value{
A list of WellLocation objects.
}
\description{
Find all wells on all plates using the following set of parameters:
}
\examples{
set1 <- findWells(contents="MTOR", experiments="brucella-du-k")
set2 <- findWells(contents=2475, experiments="brucella-au-k[1-3]")
set3 <- findWells(contents="SCRAMBLED", experiments="brucella-du-k")
}

