% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilLocalData.R
\name{getAllLocalPlates}
\alias{getAllLocalPlates}
\title{Get all local plates}
\usage{
getAllLocalPlates()
}
\value{
A list of PlateLocation objects.
}
\description{
Return plate location objects for each locally present plate.
}
\examples{
plates <- getAllLocalPlates()
}

