% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarcode.R
\name{getBarcode}
\alias{getBarcode}
\title{Get the barcode of a PlateLocation/WellLocation object}
\usage{
getBarcode(x)
}
\arguments{
\item{x}{A Data/DataLocation object}
}
\value{
The barcode of the PlateLocation/WellLocation object (as a string)
}
\description{
Given a PlateLocation/WellLocation object, return the barcode
}
\examples{
plate <- PlateLocation("J101-2C")
barcode <- getBarcode(plate)

}

