% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCompletePlateAggregate.R
\name{getCompletePlateAggregate}
\alias{getCompletePlateAggregate}
\title{Get the complete plate aggregate for a PlateLocation/WellLocation object}
\usage{
getCompletePlateAggregate(x)
}
\arguments{
\item{a}{PlateLocation/WellLocation object}
}
\value{
A list with slots for kinome (kin) and genome (gen) aggregate data
}
\description{
Given a PlateLocation/WellLocation object, return the complete and
unprocessed genome and kinome plate aggregate information.
}
\examples{
plate <- PlateLocation("J101-2C")
aggregate <- getCompletePlateAggregate(plate)
}

