% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatureNames.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{List feature names of a Data object}
\usage{
getFeatureNames(x)
}
\arguments{
\item{x}{The MatData/PlateData/WellData/ImageData object of interest}
}
\value{
A list of all currently available features
}
\description{
Given a MatData/PlateData/WellData/ImageData object, list all currently
stored features.
}
\examples{
plate    <- PlateData(PlateLocation("J101-2C"))
features <- getFeatureNames(plate)
}

