% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLinearizedPlateIndex.R
\name{getLinearizedPlateIndex}
\alias{getLinearizedPlateIndex}
\title{Linearize the index of a 16x24 well plate}
\usage{
getLinearizedPlateIndex(row, col, img = NULL)
}
\arguments{
\item{row}{Row of the well, specified either by A:P, a:P, or 1:16}

\item{col}{Column number of the well, specified by 1:24}

\item{img}{NULL/Image number within the well, 1:9}
}
\value{
The linearized index; if img==NULL, range 1:384, else range: 1:3456
}
\description{
Given row, column and possible image coordinates on a plate, get the
corresponding index of a linearized representation
}
\examples{
getLinearizedPlateIndex("B", 12, NULL)
getLinearizedPlateIndex(2, 12, 5)
}

