% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLocation.R
\name{getMetadataCacheFilename}
\alias{getMetadataCacheFilename}
\title{Get the path/filename of the metadata chache for a PlateLocation/WellLocation
object}
\usage{
getMetadataCacheFilename(x)
}
\arguments{
\item{a}{PlateLocation/WellLocation object}
}
\value{
The path/filename of the metadata chache for a
        PlateLocation/WellLocation object (as a string)
}
\description{
Given a PlateLocation/WellLocation object, return the path/filename of the
corresponding metadata chache file.
}
\examples{
plate <- PlateLocation("J101-2C")
path <- getMetadataCacheFilename(plate)
}

