% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNoImgPerWell.R
\name{getNoImgPerWell}
\alias{getNoImgPerWell}
\title{Get the barcode of a PlateLocation/WellLocation object}
\usage{
getNoImgPerWell(data)
}
\arguments{
\item{data}{A Data/DataLocation object}
}
\value{
The barcode of the PlateLocation/WellLocation object (as a string)
}
\description{
Given a PlateLocation/WellLocation object, return the barcode
}
\examples{
plate <- PlateLocation("J101-2C")
barcode <- getNoImgPerWell(plate)
}

