% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilMulticore.R
\name{getNumCores}
\alias{getNumCores}
\title{Get the number of available cores}
\usage{
getNumCores()
}
\value{
The number of available cores.
}
\description{
In case, the package is used in an LSF environment, detectCores() will
report the number of physical cores insted of the numer allocated by LSF.
}
\examples{
n.cores <- getNumCores()
}

