% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPlateWells.R
\name{getPlateWells}
\alias{getPlateWells}
\title{Fetch plate wells specified by experiment, barcode, row and column}
\usage{
getPlateWells(coordinates, force.download = FALSE, force.import = FALSE,
  keep.files = FALSE)
}
\arguments{
\item{coordinates}{Table holding locations of wells (columns for
experiment, barcode, well row and well column)}

\item{force.download}{Force downloading from openBIS}

\item{force.import}{Force importing from matlab files}

\item{keep.files}{Keep matlab files after successful import and
                      well data extraction}
}
\value{
A nested list with the top level hierarchy corresponding to a plate
        and the level below to a well and for structure below, see
        importCompletePlate
}
\description{
Given results from findPlateWellsFromGene, or a table containing columns for
experiment, barcode, well row and well column, fetch all needed plates (from
openBIS if needed, or from matlab files if present) and extract the specified
wells (result are saved such that plates do not have to be re-imported later
on)
}
\examples{
# get gene locations
mtor.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "MTOR")
scra.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "SCRAMBLED")
# combine for faster fetching
targets <- rbind(mtor.bruc.du[1:6,], scra.bruc.du[c(6,18,78,90,150,162),])

data <- getPlateWells(targets)
}

