% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getScRdsCacheFilename.R
\name{getScRdsCacheFilename}
\alias{getScRdsCacheFilename}
\title{Get path/filename of a PlateLocation/WellLocation object single cell data
cache}
\usage{
getScRdsCacheFilename(x)
}
\arguments{
\item{a}{PlateLocation/WellLocation object}
}
\value{
The path/filename of the single cell data cache of
        PlateLocation/WellLocation object (as a string)
}
\description{
Upon downloading from openBIS and importin to R, all single cell data
associated with a PlateLocation/WellLocation is cached in an .rds file. This
function gets the filepath of this cache file.
}
\examples{
plate <- PlateLocation("J101-2C")
path <- getScRdsCacheFilename(plate)
}

