% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSingleCellData.R
\name{getSingleCellData}
\alias{getSingleCellData}
\title{Read all single cell features for wells/plates}
\usage{
getSingleCellData(locations, select.features = NULL, drop.features = NULL)
}
\arguments{
\item{locations}{A list of PlateLocation/WellLocation objects for which
the data is to be returned.}

\item{select.features}{A vector of strings that are matched with features to
be kept.}

\item{drop.features}{A vector of strings that are matched with features to
                       be dropped.}
}
\value{
A list with well/plate data
}
\description{
Read all available single cell feature data for multiple wells and/or
multiple plates.
}
\examples{
plate1 <- PlateLocation("J101-2C")
mtor   <- findPlateWellsFromGene("MTOR", "brucella-du-k")
plate2 <- PlateLocation("J101-2L")

locations <- c(list(plate1, plate2), mtor)

data <- getSingleCellData(locations)
}

