% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getWellIndex.R
\name{getWellIndex1D}
\alias{getWellIndex1D}
\title{Linearize the index of a well on a 16x24 well plate}
\usage{
getWellIndex1D(row, col, img = NULL)
}
\arguments{
\item{row}{Row of the well, specified either by A:P, a:P, or 1:16}

\item{col}{Column number of the well, specified by 1:24}

\item{img}{NULL/Image number within the well, 1:9}
}
\value{
The linearized index; if img==NULL, range 1:384, else range: 1:3456
}
\description{
Given row, column and possible image coordinates on a plate, get the
corresponding index of a linearized representation
}
\examples{
ind1 <- getWellIndex1D("B", 12, NULL)
ind2 <- getWellIndex1D(2, 12, 5)
}

