% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmBootstrap.R
\name{glmBootstrapStability}
\alias{glmBootstrapStability}
\title{Analyze stability of top n coefficients}
\usage{
glmBootstrapStability(well.a, well.b, n.rep = 100, n.hit = 20,
  frac.sample = 0.7, seed = 7, glm.fun = "glmnet", norm.feat = "all",
  norm.method = "zScore", norm.sep = FALSE, n.cores = getNumCores(), ...)
}
\arguments{
\item{well.a}{A WellData object for glm analysis.}

\item{well.b}{A second WellData object for glm analysis.}

\item{n.rep}{The number of bootstrap repeats.}

\item{n.hit}{Record this many top coefficients.}

\item{frac.sample}{Size of the subsample as fraction of all available data.}

\item{seed}{Seed for reproducible sampling.}

\item{glm.fun}{The glm function to use.}

\item{n.cores}{The number of cores to use for bootstrapping.}

\item{...}{Further arguments to be passed to the glm function. For
                   example using glmnet, the alpha value.}

\item{normalize}{Whether to normalize the data (consist of discarding
separating variables and scaling intensity features to
[0,1])}
}
\value{
A vector of frequencies for features that were among the top n.hit
        coefficients.
}
\description{
Resamples the data n.rep times, fits a glm modle using the function glm.fun
and records the top n.hit coefficient names. Returns a table with
frequencies of features that were among the top n.hit coefficients.
}
\examples{
wells <- findWells(plates=c("J110-2C", "J104-2D", "J107-2L"),
                   well.names=c("H2", "H6"))
data  <- unlist(getSingleCellData(wells), recursive=FALSE)
a <- cleanData(data[["J110-2C.H2"]], "lower")
b <- cleanData(data[["J110-2C.H6"]], "lower")
test <- glmBootstrapStability(a, b, n.rep=20)
}

