% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilMulticore.R
\name{\%doparMC\%}
\alias{\%doparMC\%}
\title{A wrapper around '\%dopar\%' to recover messages/warnings in GUI environments}
\usage{
obj \%doparMC\% expr
}
\arguments{
\item{obj}{A foreach object like for regular \%dopar\% (see foreach package
documentation)}

\item{expr}{An expression to be evaluated in parallel like for regular
            \%dopar\% (see foreach package documentation)}
}
\value{
A list with slots res for the actual result of the parallel
        computation and out for the output generated by the individual
        processes
}
\description{
A fifo connection is opened and all output from the forked processes is
written to a temp file. After all processes finish, the file is read, deleted
and its contents returned together with the result of the actual parallel
computation.
}
\examples{
test <- function() {
  message("abc")
  warning("def")
  1:5 + 1:3
}

library(singleCellFeatures)

registerDoMC(cores=detectCores())
res <- foreach(1:10, .combine=rbind) \%doparMC\% test()

message(paste(res$out, collapse="\\n"))
res$res
}

