% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importCompletePlate.R
\name{importCompletePlate}
\alias{importCompletePlate}
\title{Read specified single cell features for a single plate}
\usage{
importCompletePlate(path, features = NULL)
}
\arguments{
\item{path}{Folder path to folder corresponding to a single plate,
holding single cell feature data in .mat files somewhere
below}

\item{features}{Vector holding strings corresponding to features to be
                imported}
}
\value{
A nested list; for structure see examples section
}
\description{
Read all available/the specified single cell feature data for a plate given
by its path
}
\examples{
path     <- paste("/Users/nbennett/Polybox/MasterThesis/openBISDownload/",
                  "INFECTX_PUBLISHED/ADENO_TEAM/ADENO-AU-K2/KB2-03-1J",
                  sep="")
features <- c("location")
dat.plat <- importCompletePlate(path, features)

\dontrun{
  --> structure of outplut list, representing a plate
    |--> metadata for plate (14 fields)
    |--> data (list of length 384; 24 cols by 16 rows)
       |--> A1
          |--> ...
      ...
       |--> P24
          |--> metadata for well (40 fields)
          |--> data (list of length 9; 9 images per well)
             |--> img_11
                |--> ...
            ...
             |--> img_33
                |--> Feature1
               ...
                |--> FeatureN
}
}

