% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importMultiplePlates.R
\name{importMultiplePlates}
\alias{importMultiplePlates}
\title{Read single cell feature data for multiple plates}
\usage{
importMultiplePlates(path, plates = NULL, features = NULL)
}
\arguments{
\item{path}{Folder path to folder corresponding to an experiment}

\item{plates}{Names of plates belonging to the given experiment that are
to be imported}

\item{features}{Vector holding strings corresponding to features to be
                imported}
}
\value{
A nested list with the top level hierarchy corresponding to a plate
        and the level below to a feature
}
\description{
Read all available/the specified set of single cell feature data for all
plates/the specified set below the dir path
}
\examples{
path     <- paste("/Users/nbennett/Polybox/MasterThesis/openBISDownload/",
                  "INFECTX_PUBLISHED/ADENO_TEAM", sep="")
plates   <- c("KB2-03-1I", "KB2-03-1J")
features <- c("Cells.AreaShape_Area", "Cells.AreaShape_Eccentricity")
dat.exp1 <- importMultiplePlates(path, NULL, features)
dat.exp2 <- importMultiplePlates(path, plates, features)
}

