% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSingleCellData.R
\name{importSingleCellDataPlate}
\alias{importSingleCellDataPlate}
\title{Read all single cell features for a single plate}
\usage{
importSingleCellDataPlate(plate)
}
\arguments{
\item{plate}{PlateLocation object corresponding to the plate of interest}
}
\value{
A nested list; for structure see examples section
}
\description{
Read all available single cell feature data for a plate given by a
PlateLocation object
}
\examples{
plate <- PlateLocation("J101-2C")
data  <- importSingleCellDataPlate(plate)

\dontrun{
  --> structure of outplut list a PlateData object
    |--> metadata for plate
    |--> data (list of 384 WellData objects)
       |--> A1
      ...
       |--> P24
          |--> metadata for well
          |--> data (list of 9 ImageData objects)
             |--> img_11
            ...
             |--> img_33
}
}

