% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeFeatureCompatible.R
\name{makeFeatureCompatible}
\alias{makeFeatureCompatible}
\title{Enforce equal feature sets}
\usage{
makeFeatureCompatible(lst)
}
\arguments{
\item{lst}{A list of DataObjects.}
}
\value{
A list of DataObjects (same type as input) which is guaranteed to
        have the same feature set.
}
\description{
Given a list of DataObject, ensure that all list enstries have the same set
of features.
}
\examples{
wells <- findWells(plates=c("J110-2C", "J104-2D", "J107-2L"),
                   well.names=c("H2", "H6"))
data  <- unlist(getSingleCellData(wells), recursive=FALSE)
data  <- lapply(data, cleanData, "lower")
data  <- makeFeatureCompatible(data)
}

