% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareDataforGlm.R
\name{makeRankFull}
\alias{makeRankFull}
\title{Enforce full rank design matrix}
\usage{
makeRankFull(data)
}
\arguments{
\item{data}{Matrix/data frame holding the design matrix and optionally a
               response vector named "Response"}
}
\value{
A subset of the original data structure (some of the variables are
        removed)
}
\description{
Given a design matrix (if the response is also included it has to be named
"Response"), rank deficiency is corrected by dropping zero variance variables
and one of each pair of highly correlated variables to make the design matrix
invertible.
}
\examples{
# get gene locations
mtor.loc <- findWells(experiments="brucella-du-k1", contents="MTOR")
scra.loc <- findWells(plates=sapply(mtor.loc, getBarcode),
                      contents="SCRAMBLED", well.names="G23")
# combine for faster fetching
data     <- getSingleCellData(list(mtor.loc[[1]], scra.loc[[1]]))
mtor.dat <- meltData(cleanData(data[[1]]$H6))
scra.dat <- meltData(cleanData(data[[1]]$G23))
# prepare data for glm
data <- prepareDataforGlm(mtor.dat$mat$Cells, scra.dat$mat$Cells)
data <- makeRankFull(data)
# run glm (enforcing full rank might be advisbale)
model <- glm("Response ~ .", binomial, data$train)
}

