% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{MatData}
\alias{MatData}
\title{Constructor for MatData objects}
\usage{
MatData(plate, force.download = FALSE)
}
\arguments{
\item{plate}{PlateLocation object corresponding to the plate of interest}
}
\value{
A nested list; for structure see examples section
}
\description{
Given a PlateLocation object, fetch (from file or from openBIS) the
corresponding MatData object.
}
\examples{
plate <- PlateLocation("J101-2C")
data  <- MatData(plate)

\dontrun{
  --> structure of output
    |--> meta
       |--> metadata for plate
    |--> data
       |--> Feature_1
      ...
       |--> Feature_3456
}

}

