% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meltData.R
\name{meltData}
\alias{meltData}
\title{Combine all data in a given Data object}
\usage{
meltData(x)
}
\arguments{
\item{x}{The MatData/PlateData/WellData/ImageData object of interest.}
}
\value{
A list of data frames for the combined objects.
}
\description{
Given a PlateData/WellData/ImageData object, rbind all data contained in the
hierarchy below.
}
\examples{
plate    <- PlateData(PlateLocation("J101-2C"))
plate.df <- meltData(plate)
}

