% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meltData.R
\name{moveFeatures}
\alias{moveFeatures}
\title{Move feature(s) to different group}
\usage{
moveFeatures(data, from, to, aggregate = "mean")
}
\arguments{
\item{data}{The data (the result of a call to meltData) to be acted
upon.}

\item{from}{The feature name (or vector of feature names) to be added.}

\item{to}{The group to add them to.}

\item{aggregate}{If the target group is smaller, this function is used to
                 aggregate the data.}
}
\value{
The data.frame specified with the parameter to with the chosen
        features added.
}
\description{
A specified feature (or a vector of features) is moved to the desired group.
If the target group has more rows than the source group, the features are
replicated according to the respective pattern and if the target group has
fewer rows, the feature is summarized with the supplied function.
}
\examples{
data   <- PlateData(PlateLocation("J107-2C"))
molten <- meltData(data)
test   <- moveFeatures(molten, from="Cells.AreaShape_Area", to="Image")
}

