% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareDataforGlm.R
\name{normalizeMoltenData}
\alias{normalizeMoltenData}
\title{Normalize data}
\usage{
normalizeMoltenData(data, features = "all", method = "zScore",
  separate.wells = FALSE)
}
\arguments{
\item{data}{The data frame holding the data to be normalized.}

\item{features}{A regular expression specifying the features to be
normalized. The string "all" includes all available.}

\item{method}{A string specifying the normalization method to be
used. The following are implemented: unitInterval,
scaleRange, center, unitVar, zScore.}

\item{separate.wells}{Are the two wells to be treated seperately or
                      concurrently?}
}
\value{
A data.frame holding normalized data.
}
\description{
Given two well level matrices combined by prepareDataforGlm, normalize the
data.
}
\examples{
wells     <- findWells(plates="J107-2C", well.names=c("H2", "H6"))
data      <- unlist(getSingleCellData(wells), recursive=FALSE)
cleaned   <- lapply(data, cleanData, "lower")
melted    <- lapply(cleaned, meltData)
data.mint <- prepareDataforGlm(melted[[2]]$mat$Cells,
                               melted[[1]]$mat$Cells, test=NULL)$train
data.scal <- normalizeMoltenData(data.mint, features="intensity",
                                 method="unitInterval")
}

