% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePlateData.R
\name{plateBoxplot}
\alias{plateBoxplot}
\title{Boxpolot of feature per well}
\usage{
plateBoxplot(plate.dat, feature)
}
\arguments{
\item{plate.dat}{Data used to generate boxplot Expecting a PlateData
object.}

\item{feature}{Name of the feature to plot. Expecting a string.}
}
\value{
A ggplot2 plot object ready for printing.
}
\description{
Given plate data and a feature name, create a boxplot of individual wells,
faceted by row indices.
}
\examples{
plate <- PlateLocation("J101-2C")
data  <- PlateData(plate)
plateBoxplot(data, "Cells.AreaShape_Area")

}

