% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePlateData.R
\name{plateHeatmap}
\alias{plateHeatmap}
\title{Plot a plate heatmap}
\usage{
plateHeatmap(plate.dat, feature, fun.aggregate = "mean",
  fun.transform = "identity", show.control = TRUE, colors = brewer.pal(9,
  "Reds")[3:9])
}
\arguments{
\item{plate.dat}{Data used to generate heatmap. Expecting a PlateData
object.}

\item{feature}{Name of the feature to extract and aggreagte within
wells for heatmap plot. Expecting a string.}

\item{fun.aggregate}{The function used to aggregate data within wells.
Expecting either a function object or a string.}

\item{fun.transform}{An optional function used to transform the data before
plotting. Expecting either a function object or a
string.}

\item{show.control}{Logical, whether to mark the control wells.}

\item{colors}{A vector of colors that is interpolated for the 
gradient scale.}
}
\value{
A ggplot2 plot object ready for printing.
}
\description{
Given plate data, plot a heatmap of the specified feature using a function
to aggregate data within wells and an optional function to transform the
resulting vector of length 384.
}
\examples{
plate <- PlateLocation("J101-2C")
data  <- PlateData(plate)
plateHeatmap(data, "VoronoiCells.Location_Center_Y", median, sqrt)

}

