% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLocation.R
\name{PlateLocation}
\alias{PlateLocation}
\title{Constructor for PlateLocation objects}
\usage{
PlateLocation(barcode)
}
\arguments{
\item{barcode}{A plate barcode as a character string}
}
\value{
A PlateLocation object: a list with slots plate, space, group and
        experiment
}
\description{
Given a plate specification (consisting of a plate barcode) an S3 object of
type PlateLocation is created
}
\examples{
plate <- PlateLocation("J107-2C")
}

