% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepareDataforGlm.R
\name{prepareDataforGlm}
\alias{prepareDataforGlm}
\title{Prepare a dataset for GLM analysis}
\usage{
prepareDataforGlm(active, control)
}
\arguments{
\item{active}{Data coming from a well where a gene knockdown occurred}

\item{control}{Data belonging to a control well}
}
\value{
A list with entries "test" and "train" each holding a data frame
        containing a design matrix and a response vector
}
\description{
Given two design matrices (one corresponding to a knockdown and one to a
control well), create a concatenated design matrix with a column "Response"
encoding for original membership. This unified data matrix is then divided
into 90% training and 10% testing. To avoid missingness problems, all
variables containing NA/NaN are dropped.
}
\examples{
# get gene locations
mtor.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "MTOR")
scra.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "SCRAMBLED")
# combine for faster fetching
targets <- rbind(mtor.bruc.du[1:6,], scra.bruc.du[c(6,18,78,90,150,162),])
# fetch data
data <- getPlateWells(targets)
# prepare data for glm
j101_2c <- prepareDataforGlm(data[["J101-2C"]]$H6$data,
                             data[["J101-2C"]]$G23$data)
# run glm (enforcing full rank might be advisbale)
model <- glm("Response ~ .", binomial, j101_2c$train)
}

