% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPigz.R
\name{readRDSMC}
\alias{readRDSMC}
\title{A wrapper around readRDS for faster reading of compressed files}
\usage{
readRDSMC(file)
}
\arguments{
\item{file}{The file to be read}
}
\value{
The object saved as file
}
\description{
Although decompressing files with pigz is not parallelized, it is faster to
run the decompression in a separate process.
}
\examples{
data <- readRDSMC("~/obj.rds")
}

