% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilLocalData.R
\name{rebuildAllMatDataCaches}
\alias{rebuildAllMatDataCaches}
\title{Rebuild all local plates}
\usage{
rebuildAllMatDataCaches()
}
\value{
A list booleans indicating whether the plate was successfully
        rebuilt. Rebuilding complete plates counts as failure.
}
\description{
For all local plates, retry downloading missing features and rebuild the
corresponding caches.
}
\examples{
res <- rebuildAllMatDataCaches()
}

