% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{reloadSingleCellFeatures}
\alias{reloadSingleCellFeatures}
\title{Get directory of the package source}
\usage{
reloadSingleCellFeatures()
}
\value{
The package source directory as a string
}
\description{
Returns the directory of the package source
}
\examples{
dir <- getPackageDir()
}

