% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPigz.R
\name{saveRDSMC}
\alias{saveRDSMC}
\title{A wrapper around saveRDS to enable multicore compression}
\usage{
saveRDSMC(object, file, threads = getNumCores())
}
\arguments{
\item{object}{An object to be saved}

\item{file}{The filename of the resulting .rds file}

\item{threads}{the number of threads to be used for compression (default:
all-1)}
}
\value{
NULL (invisibly). The object is saved to the filesystem.
}
\description{
Piping the datastream to be saved through pigz enables the use of multiple
cores for compression.
}
\examples{
obj <- list(a=1, b=2)
saveRDSMC(obj, "~/obj.rds")

}

