% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveToCache.R
\name{saveToCache}
\alias{saveToCache}
\title{Build cache files for certain Data objects}
\usage{
saveToCache(x, force.write = FALSE)
}
\arguments{
\item{x}{MatData/WellData object}

\item{force.write}{Logical value indicating whether the cache file is
                   overwritten if already present}
}
\value{
NULL (invisibly). The cache file is saves to its corresponding
        location.
}
\description{
Given a MatData/WellData object, save it to the corresponding cache location
}
\examples{
plate <- PlateData(PlateLocation("J101-2C"))
well  <- extractWell(plate, WellLocation(plate, "A", 5), FALSE)
saveToCache(well)
}

