% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilDatabases.R
\name{updateDatabaseFeatures}
\alias{updateDatabaseFeatures}
\title{Update the featureDatabase object}
\usage{
updateDatabaseFeatures()
}
\value{
NULL (invisibly). The updated featureDatabase object saved to the
        /data folder. For the new file to be available, the package has to
        be reloaded.
}
\description{
In order to update the featureDatabase object, all *.txt files in the
folder /path/to/meta.dir/FeatureLists are read, processed and the feature
"Cells.Infection_IsInfected" is added. The filenames determine the list slot
names, so they should be formatted as PATHOGEN-*.txt.
}
\examples{
updateDatabaseFeatures()

}

