% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilDatabases.R
\name{updateDatabasePlate}
\alias{updateDatabasePlate}
\title{Generate the plateDatabase object}
\usage{
updateDatabasePlate()
}
\value{
NULL, invisibly. The updated plateDatabase object is saved to the
        /data folder. For  the new file to be available, the package has to
        be reloaded.
}
\description{
Generate the plateDatabase object (holding all neccessary information to
locate any plate on openBIS/in a local directory structure created by
BeeDataSetDownloader) using genome aggregate files.
}
\examples{
updateDatabasePlate()
# using the plateDatabase object
data(plateDatabase)
str(plate.database)

}

