% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateDatabases.R
\name{updateDatabaseSettingsGet}
\alias{updateDatabaseSettingsGet}
\title{Retrieve the settingsDatabase object}
\usage{
updateDatabaseSettingsGet()
}
\value{
The current settingsDatabase object
}
\description{
In order to update the settingsDatabase object, this helper function fetches
the current settingsDatabase object for the user to modify and save using
updateDatabaseSettingsSet
}
\examples{
# load current settingsDatabase
current <- updateDatabaseSettingsGet()
# save updated version
updateDatabaseSettingsSet(current, list(newKey="newValue"))
# using the settingsDatabase object
data(settingsDatabase)
str(settings.database)
# just updating the settingsDatabase object
updateDatabaseSettingsSet(updateDatabaseSettingsGet())
}

