% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilDatabases.R
\name{updateDatabaseSettingsSet}
\alias{updateDatabaseSettingsSet}
\title{Save the settingsDatabase object}
\usage{
updateDatabaseSettingsSet(old.list, new.list = NULL)
}
\arguments{
\item{old.list}{Required parameter holding the old list/a new version of
settingsDatabase with e.g. a key-value pair removed}

\item{new.list}{Optional parameter holding one or more key-value pairs as a
                list which will be appended to old.list}
}
\value{
The updated settingsDatabase object which is also saved to the /data
        folder. For the new file to be available, the package has to be re-
        loaded
}
\description{
In order to update the settingsDatabase object, updateDatabaseSettingsGet
can be used to fetch the current settingsDatabase object, which then can be
modifyed and saved using this function
}
\examples{
# load current settingsDatabase
current <- updateDatabaseSettingsGet()
# save updated version
updateDatabaseSettingsSet(current, list(newKey="newValue"))
# using the settingsDatabase object
data(settingsDatabase)
str(settings.database)
# just updating the settingsDatabase object
updateDatabaseSettingsSet(updateDatabaseSettingsGet())
}

