% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilDatabases.R
\name{updateDatabaseWells}
\alias{updateDatabaseWells}
\title{Generate the wellDatabase{Pathogen} objects}
\usage{
updateDatabaseWells(pathogens = NULL)
}
\arguments{
\item{pathogens}{An optional list of pathogen names to have their
                 corresponding database updated. If not specified, all
                 pathogens are updated.}
}
\value{
NULL, invisibly. The updated plateDatabase object for each of the
        specified pathogens is saved to the /data folder. For the new files
        to be available, the package has to be reloaded.
}
\description{
For each specified pathogen, generates the corresponding
wellDatabase{Pathogen} object (holding well type, target gene ID and name for
each well) using genome aggregate files, as well as a kinome aggregate file
which contains additional information of control types which are not
currently available for genome plates.
}
\examples{
updateDatabaseWells(c("adeno", "brucella"))
# using the wellDatabase{Pathogen} objects
data(wellDatabaseAdeno)
str(well.database.adeno)
}

