% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilDatabases.R
\name{wellDatabaseCoverage}
\alias{wellDatabaseCoverage}
\title{Assess the coverage provided by available metadata}
\usage{
wellDatabaseCoverage(verbose = FALSE, is.startup = FALSE)
}
\arguments{
\item{verbose}{A verbosity argument, specifying whether to print all
plates that are missing of just a summary.}

\item{is.startup}{Only relevant when this function is used upon package
                  loading to output via packageStartupMessage.}
}
\value{
NULL, invisibly. All information is printed.
}
\description{
The plate database is compared to all well databases in order to find all
plates that are missing in well databases. The plate database is assumed
to be complete in the sense that it is generated through a data set search
in openBIS and should therefore include all plates for which single cell
data is available. The well databases however are extracted from aggregates
which only become available periodically.
}
\examples{
wellDatabaseCoverage()
wellDatabaseCoverage(TRUE)
}

