% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLocation.R
\name{WellLocation}
\alias{WellLocation}
\title{Constructor for WellLocation objects}
\usage{
WellLocation(plate, row, col)
}
\arguments{
\item{plate}{A plate barcode as a character string}

\item{row}{Row specification (a:p, A:P, or 1:16)}

\item{col}{Column specification (1:24)}
}
\value{
A WellLocation object: a list with slots plate, row, column, index
        (the linearized well index), space, group and experiment
}
\description{
Given a well specification consisting of plate barcode, row and column, an
S3 object of type WellLocation is created
}
\examples{
well <- WellLocation("J107-2C", "H", 6)
}

