% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Metadata.R
\name{WellMetadata}
\alias{WellMetadata}
\title{Constructor for WellMetadata objects}
\usage{
WellMetadata(well)
}
\arguments{
\item{well}{A WellLocation object}
}
\value{
A WellMetadata object: a list with slots describing conditions in
        a well.
}
\description{
Given a WellLocation object, the corresponding metadata is compiled
}
\examples{
meta <- WellMetadata(WellLocation("J107-2C", "H", 6))
}

