% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentBscore.R
\name{augmentBscore}
\alias{augmentBscore}
\title{Augment PlateData with B-scores}
\usage{
augmentBscore(x, ...)
}
\arguments{
\item{x}{The PlateData object of interest.}

\item{features}{A (list) of regular expressions used for selecting the
features to include.}

\item{drop}{A (list) of regular expressions used for removing features
from the previously compiled list.}

\item{func.aggr}{The function used to aggregate data within wells.}
}
\value{
An augmented version of the input object.
}
\description{
Given a PlateData, aggregate the selected features within wells and
calculate their plate B-scores.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
aug  <- augmentBscore(data)

}

