% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/augmentCordinateFeatures.R
\name{augmentCordinateFeatures}
\alias{augmentCordinateFeatures}
\title{Augment all features involving locations}
\usage{
augmentCordinateFeatures(x, ...)
}
\arguments{
\item{x}{The PlateData/WellData/ImageData object of interest.}

\item{ellipse}{An integer corresponding to the number of ellipses.}

\item{facet}{A vector of length 2 e.g. c(14, 10) corresponding to the
number of bins in x-direction (14) and y-direction (10).}

\item{center.dist}{Logical, whether to include distances to image/well
centers.}

\item{density}{Logical, whether to estimate densities.}
}
\value{
The augmented version of the original object.
}
\description{
All features containing the strings "Location_Center_X" or
"Location_Center_Y" are extended to include their membership to an ellipse (
concentric w.r.t. image center/well center), their membership to a tile (at
well and image level), the number of nonempty neighbor tiles, the type of
tile (inwards/outwards facing image border/edge) the distance to the image/
well center, the local object density to the image center. If features
labelled  with "Location_Shifted_X" or "Location_Shifted_Y" are present
(augmentImageLocation), Those are extended too.
}
\examples{
data <- PlateData(PlateLocation("J101-2C"))
aug <- augmentCordinateFeatures(data, 5, c(14, 10), TRUE, TRUE)
}

