% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractImages.R
\name{extractImages}
\alias{extractImages}
\title{Extract data for images from Data objects}
\usage{
extractImages(x, images, ...)
}
\arguments{
\item{x}{The original Data object, holding the superset of images}

\item{images}{A vector of integers, specifying image indices of the images
to keep}

\item{keep.well}{(Only at WellData level) A boolean indicating whether to keep
                 the WellData object intact or return ImageData object(s)}
}
\value{
Either a PlateData, WellData or ImageData object holding only the
        data corresponding to the specified subset of images.
}
\description{
Given a Data object and a vector of image indices (either in 1:6 or in 1:9)
return only the data belonging to the selected images. If this extraction is
performed on well level, either a well with only a subset of images is
returned or a single ImageData object (or a list of ImageData) objects is/are
returned.
}
\examples{
plate  <- PlateData(PlateLocation("J101-2C"))
img_11 <- extractImages(plate, 1)
}

