% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findPlateWellsFromGene.R
\name{findPlateWellsFromGene}
\alias{findPlateWellsFromGene}
\title{Find wells corresponding to a gene}
\usage{
findPlateWellsFromGene(gene, experiment, plates = NULL, verbose = FALSE)
}
\arguments{
\item{gene}{String/integer specifying the gene (case sensitive)}

\item{experiment}{Name of the experiment to be considered (should start with
pathogen name rest is optional and used for narrowing down
the number of results, all converted to uppercase)}

\item{plates}{An optional vector of plate names to further restrict the
results}

\item{verbose}{Verbosity argument (FALSE/TRUE)}
}
\value{
A table containing columns for experiment, barcode, well row and well
        column
}
\description{
Find all wells on all plates of a given pathogen/experiment that contain
the specified gene
}
\examples{
set1 <- findPlateWellsFromGene("MTOR", "brucella-du-k")
set2 <- findPlateWellsFromGene(2475, "brucella-au-k[1-3]")
set3 <- findPlateWellsFromGene("SCRAMBLED", "brucella-du-k")
}

