% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBarcode.R
\name{getBarcode}
\alias{getBarcode}
\title{Get the barcode of a PlateLocation/WellLocation object}
\usage{
getBarcode(x)
}
\arguments{
\item{a}{PlateLocation/WellLocation object}
}
\value{
The barcode of the PlateLocation/WellLocation object (as a string)
}
\description{
Give a PlateLocation/WellLocation object, return the barcode
}
\examples{
plate <- PlateLocation("J101-2C")
barcode <- getBarcode(plate)
}

