% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNoImgPerWell.R
\name{getNoImgPerWell}
\alias{getNoImgPerWell}
\title{Get the number of images per Well}
\usage{
getNoImgPerWell(data)
}
\arguments{
\item{data}{A Data (MatData/ImageData/WellData/PlateData) object}
}
\value{
The number of images per well (either 6 or 9).
}
\description{
Given a Data (MatData/ImageData/WellData/PlateData) object, return the number
of images per well (either 6 or 9).
}
\examples{
plate <- PlateData()PlateLocation("J101-2C"))
n.img <- getNoImgPerWell(plate)
}

