% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLocation.R
\name{getOpenBisPath}
\alias{getOpenBisPath}
\title{Get the path of a PlateLocation/WellLocation object on openBIS}
\usage{
getOpenBisPath(x)
}
\arguments{
\item{a}{PlateLocation/WellLocation object}
}
\value{
The openBIS location path of the PlateLocation/WellLocation object
        (as a string)
}
\description{
Given a PlateLocation/WellLocation object, return the openBIS location path
}
\examples{
plate <- PlateLocation("J101-2C")
path <- getOpenBisPath(plate)
}

