% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleCellData.R
\name{getSingleCellData}
\alias{getSingleCellData}
\title{Read all single cell features for wells/plates}
\usage{
getSingleCellData(locations, select.features = NULL, drop.features = NULL,
  select.images = NULL)
}
\arguments{
\item{locations}{A list of PlateLocation/WellLocation objects for
which the data is to be returned.}

\item{select.features}{A vector of strings that are matched with features to
be kept.}

\item{drop.features}{A vector of strings that are matched with features to
be dropped.}

\item{select.images}{A vector of image indices (1:9) indicating which
images to return for each well.}
}
\value{
If a single PlateLocation/WellLocation is specified as location 
        input, the corresponding Data object is returned. For a list of
        PlateLocation/WellLocation objects, a list with well/plate data is
        returned.
}
\description{
Read all available single cell feature data for multiple wells and/or
multiple plates.
}
\examples{
plate1 <- PlateLocation("J101-2C")
mtor   <- findPlateWellsFromGene("MTOR", "brucella-dp-k[1-3]")
plate2 <- PlateLocation("J101-2L")

locations <- c(list(plate1, plate2), mtor)

data <- getSingleCellData(locations)

}

