% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLocation.R
\name{getWellName}
\alias{getWellName}
\title{Get the well name of a WellLocation object}
\usage{
getWellName(x)
}
\arguments{
\item{A}{WellLocation object}
}
\value{
The well name of the WellLocation object (as a string)
}
\description{
Given a WellLocation object, return the well name
}
\examples{
plate <- WellLocation("J101-2C", "B", 15)
well.name <- getWellName(plate)
}

