% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeRankFull.R
\name{makeRankFull}
\alias{makeRankFull}
\title{Enforce full rank design matrix}
\usage{
makeRankFull(data)
}
\arguments{
\item{data}{Matrix/data frame holding the design matrix and optionally a
            response vector named "Response"}
}
\value{
A subset of the original data structure (some of the variables are
        removed)
}
\description{
Given a design matrix (if the response is also included it has to be named
"Response"), rank deficiency is corrected by dropping zero variance variables
and one of each pair of highly correlated variables to make the design matrix
invertible.
}
\examples{
# get gene locations
mtor.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "MTOR")
scra.bruc.du <- findPlateWellsFromGene("brucella", "du-k", "SCRAMBLED")
# combine for faster fetching
targets <- rbind(mtor.bruc.du[1:6,], scra.bruc.du[c(6,18,78,90,150,162),])
# fetch data
data <- getPlateWells(targets)
# prepare data for glm
j101_2c <- prepareDataforGlm(data[["J101-2C"]]$H6$data,
                             data[["J101-2C"]]$G23$data)
# enforce full rank
training <- makeRankFull(j101_2c$train)
# run glm
model <- glm("Response ~ .", binomial, training)
}

