% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataMeta.R
\name{PlateAggregate}
\alias{PlateAggregate}
\title{Constructor for PlateAggregate objects}
\usage{
PlateAggregate(dl)
}
\arguments{
\item{dl}{A PlateLocation/WellLocation (DataLocation) object}
}
\value{
A list with slots for PlateLocation (plate) and aggregate data
       (data).
}
\description{
Given a PlateLocation/WellLocation object, return the complete and
unprocessed plate aggregate information.
}
\examples{
plate <- PlateLocation("J101-2C")
aggregate <- PlateAggregate(plate)
}

