% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataMeta.R
\name{PlateMetadata}
\alias{PlateMetadata}
\title{Constructor for PlateMetadata objects}
\usage{
PlateMetadata(plate)
}
\arguments{
\item{well}{A PlateLocation object}
}
\value{
A PlateMetadata object: a list with slots describing conditions on
        a plate.
}
\description{
Given a PlateLocation object, the corresponding metadata is compiled
}
\examples{
meta <- PlateMetadata(PlateLocation("J107-2C"))
}

