% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareDataforGlm.R
\name{prepareDataforGlm}
\alias{prepareDataforGlm}
\title{Prepare a dataset for GLM analysis}
\usage{
prepareDataforGlm(active, control, drop.feat = NULL, drop.sep = FALSE,
  scale = "none", test = 10)
}
\arguments{
\item{active}{Data coming from a well where a gene knockdown occurred.}

\item{control}{Data belonging to a control well.}

\item{drop.feat}{A vector of strings of column names that will be dropped.}

\item{drop.sep}{Drop variables that separate data.}

\item{scale}{Scale all matched features to the unit interval.}

\item{test}{The fraction of rows to be used for testing is 1/test. If
                 NULL is supplied, all data is used for training.}
}
\value{
A list with entries "test" and "train" each holding a data frame
        containing a design matrix and a response vector.
}
\description{
Given two design matrices (one corresponding to a knockdown and one to a
control well), create a concatenated design matrix with a column "Response"
encoding for original membership. This unified data matrix is then divided
into 90% training and 10% testing. To avoid missingness problems, all
variables containing NA/NaN are dropped.
}
\examples{
# get gene locations
mtor.loc <- findWells(experiments="brucella-du-k1", contents="MTOR")
scra.loc <- findWells(plates=sapply(mtor.loc, getBarcode),
                      contents="SCRAMBLED", well.names="G23")
# combine for faster fetching
data     <- getSingleCellData(list(mtor.loc[[1]], scra.loc[[1]]))
mtor.dat <- meltData(cleanData(data[[1]]$H6))
scra.dat <- meltData(cleanData(data[[1]]$G23))
# prepare data for glm
data <- prepareDataforGlm(mtor.dat$mat$Cells, scra.dat$mat$Cells)
data <- makeRankFull(data)
# run glm
model <- glm("Response ~ .", binomial, data$train)
}

