% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rebuildCache.R
\name{rebuildCache}
\alias{rebuildCache}
\title{Rebuild cache files for certain Data objects}
\usage{
rebuildCache(x)
}
\arguments{
\item{x}{MatData object}
}
\value{
The updated MatData object. The updated cache file is saved to its
        corresponding location.
}
\description{
Given an existing MatData cache file and some feature files that were not
imported correctly, update and save it to the corresponding cache location.
}
\examples{
data <- MatData(PlateLocation("J101-2C"))
rebuildCache(data)
}

