% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilPackage.R
\name{reloadSingleCellFeatures}
\alias{reloadSingleCellFeatures}
\title{Reload the singleCellFeatures package}
\usage{
reloadSingleCellFeatures()
}
\value{
NULL (invisibly).
}
\description{
Detaches and reloads the singleCellFeatures package.
}
\examples{
reloadSingleCellFeatures()

}

